0613 문의 상세보기 만드는중

+++ src/main/java/yep/greenFire/greenfirebackend/inquiry/site/domain/repository/SiteInquiryRepository.java
@@ -12,5 +12,5 @@ import yep.greenFire.greenfirebackend.inquiry.site.dto.response.InquiryResponse;
 public interface SiteInquiryRepository extends JpaRepository<InquiryContent, Integer> {
     Page<InquiryContent> findByMemberCode(Long memberCode, Pageable pageable);
 
-    InquiryResponse findByInquiryCode(Long inquiryCode);
+    InquiryResponse findByInquiryCode(Integer inquiryCode);
 }
